Consistency and reuse.
