+++ tests/Aspire.Hosting.Tests/Utils/UnixSocketHelper.cs
@@ -1,0 +1,23 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+namespace Aspire.Hosting.Tests.Utils;
+
+internal static class UnixSocketHelper
+{
+    public static string GetBackchannelSocketPath()
+    {
+        var homeDirectory = Environment.GetFolderPath(Environment.SpecialFolder.UserProfile);
+        var dotnetCliPath = Path.Combine(homeDirectory, ".dotnet", "aspire", "cli", "backchannels");
+
+        if (!Directory.Exists(dotnetCliPath))
+        {
+            Directory.CreateDirectory(dotnetCliPath);
+        }
+
+        var uniqueSocketPathSegment = Guid.NewGuid().ToString("N");
+        var socketPath = Path.Combine(dotnetCliPath, $"cli.sock.{uniqueSocketPathSegment}");
+        return socketPath;
+    }
+
+}
